refactor(lex-babel): split lex-format serializer into numbering/tables submodules
