+++ crates/lex-babel/src/formats/lex/serializer.rs
@@ -5,76 +5,26 @@ use lex_core::lex::ast::{
         verbatim::VerbatimGroupItemRef, VerbatimLine,
     },
     traits::{AstNode, Visitor},
-    Annotation, Definition, Document, List, ListItem, Paragraph, Session, Table, TableCell,
-    TableCellAlignment, TableRow, Verbatim,
+    Annotation, Definition, Document, List, ListItem, Paragraph, Session, Table, Verbatim,
 };
 
 use lex_core::lex::assembling::stages::normalize_labels::source_spelling;
 use lex_core::lex::ast::elements::sequence_marker::DecorationStyle;
+
+mod numbering;
+mod tables;
+
+use numbering::format_marker_index;
+use tables::emit_pipe_table;
+
+#[cfg(test)]
+mod tests;
 
 struct ListContext {
     index: usize,
     style: DecorationStyle,
     upper_case: bool,
     marker_form: Option<Form>,
-}
-
-fn format_marker_index(style: DecorationStyle, upper_case: bool, index: usize) -> String {
-    match style {
-        DecorationStyle::Plain => "-".to_string(),
-        DecorationStyle::Numerical => index.to_string(),
-        DecorationStyle::Alphabetical => {
-            if upper_case {
-                to_alpha_upper(index)
-            } else {
-                to_alpha_lower(index)
-            }
-        }
-        DecorationStyle::Roman => to_roman_upper(index),
-    }
-}
-
-fn to_alpha_lower(n: usize) -> String {
-    if (1..=26).contains(&n) {
-        char::from_u32((n as u32) + 96).unwrap().to_string()
-    } else {
-        n.to_string()
-    }
-}
-fn to_alpha_upper(n: usize) -> String {
-    if (1..=26).contains(&n) {
-        char::from_u32((n as u32) + 64).unwrap().to_string()
-    } else {
-        n.to_string()
-    }
-}
-
-fn to_roman_upper(n: usize) -> String {
-    // Convert to Roman numerals (uppercase) for common values
-    // Falls back to decimal for values > 20
-    match n {
-        1 => "I".to_string(),
-        2 => "II".to_string(),
-        3 => "III".to_string(),
-        4 => "IV".to_string(),
-        5 => "V".to_string(),
-        6 => "VI".to_string(),
-        7 => "VII".to_string(),
-        8 => "VIII".to_string(),
-        9 => "IX".to_string(),
-        10 => "X".to_string(),
-        11 => "XI".to_string(),
-        12 => "XII".to_string(),
-        13 => "XIII".to_string(),
-        14 => "XIV".to_string(),
-        15 => "XV".to_string(),
-        16 => "XVI".to_string(),
-        17 => "XVII".to_string(),
-        18 => "XVIII".to_string(),
-        19 => "XIX".to_string(),
-        20 => "XX".to_string(),
-        _ => n.to_string(), // Fallback to decimal for larger numbers
-    }
 }
 
 pub struct LexSerializer {
@@ -148,7 +98,7 @@ impl LexSerializer {
         self.rules.indent_string.repeat(self.indent_level)
     }
 
-    fn write_line(&mut self, text: &str) {
+    pub(super) fn write_line(&mut self, text: &str) {
         if self.suppress_output > 0 {
             return;
         }
@@ -480,753 +430,5 @@ impl Visitor for LexSerializer {
 
     fn leave_table(&mut self, _table: &Table) {
         // No-op; annotations carry the closer.
-    }
-}
-
-/// What occupies a single grid column in a single row.
-enum Slot<'a> {
-    /// The top-left (originating) cell of a span; carries its content.
-    Origin(&'a TableCell),
-    /// A column absorbed by a colspan to its left — re-emits `>>` (lex#683).
-    Colspan,
-    /// A column absorbed by a rowspan above it — re-emits `^^` (lex#694).
-    Rowspan,
-    /// Padding for a column a short row never reaches but that sits *before* a
-    /// rowspan-covered column further right; keeps later `^^` markers in place.
-    Empty,
-}
-
-impl Slot<'_> {
-    /// The literal a slot renders as, for width and emission.
-    fn text(&self) -> &str {
-        match self {
-            Slot::Origin(cell) => cell.content.as_string().trim(),
-            Slot::Colspan => ">>",
-            Slot::Rowspan => "^^",
-            Slot::Empty => "",
-        }
-    }
-}
-
-/// Project the table's ragged rows onto a rectangular column grid, re-deriving
-/// the merge markers the parser consumed: `>>` for each column a colspan
-/// absorbed (within a row) and `^^` for each column a rowspan absorbed (from the
-/// row above). The parser removes absorbed cells and bumps the spanning cell's
-/// colspan/rowspan, so without this reconstruction the markers — and the spans —
-/// are lost on a re-format (lex#683, lex#694).
-fn build_grid<'a>(rows: &[&'a TableRow]) -> Vec<Vec<Slot<'a>>> {
-    // `carry[col]` = remaining continuation rows still covered by a rowspan that
-    // originated above. Grows on demand since the column count emerges here.
-    let mut carry: Vec<usize> = Vec::new();
-    let mut grid: Vec<Vec<Slot>> = Vec::with_capacity(rows.len());
-
-    for row in rows {
-        let mut slots: Vec<Slot> = Vec::new();
-        let mut cells = row.cells.iter();
-        let mut col = 0;
-        loop {
-            if carry.get(col).copied().unwrap_or(0) > 0 {
-                // This column is mid-rowspan from a cell above.
-                carry[col] -= 1;
-                slots.push(Slot::Rowspan);
-                col += 1;
-            } else if let Some(cell) = cells.next() {
-                let span = cell.colspan.max(1);
-                if cell.rowspan > 1 {
-                    // Reserve the continuation rows across the cell's full width.
-                    for c in col..col + span {
-                        if carry.len() <= c {
-                            carry.resize(c + 1, 0);
-                        }
-                        carry[c] += cell.rowspan - 1;
-                    }
-                }
-                slots.push(Slot::Origin(cell));
-                for _ in 1..span {
-                    slots.push(Slot::Colspan);
-                }
-                col += span;
-            } else if carry.iter().skip(col).any(|&r| r > 0) {
-                // Cells are exhausted but a rowspan still covers a column further
-                // right; pad this hole so that column's `^^` is still emitted and
-                // its `carry` is consumed in *this* row rather than leaking down.
-                slots.push(Slot::Empty);
-                col += 1;
-            } else {
-                // No more cells and nothing covered ahead: row is done.
-                break;
-            }
-        }
-        grid.push(slots);
-    }
-    grid
-}
-
-/// Emit a structural Table as a markdown-style pipe table, padded for
-/// column alignment. The column count is the max-width row; shorter
-/// rows pad with empty cells. Alignment follows the per-cell `align`
-/// attribute, which the parser sets from the markdown alignment row
-/// (`:---`, `:---:`, `---:`).
-fn emit_pipe_table(serializer: &mut LexSerializer, table: &Table) {
-    let all_rows: Vec<&TableRow> = table
-        .header_rows
-        .iter()
-        .chain(table.body_rows.iter())
-        .collect();
-    if all_rows.is_empty() {
-        return;
-    }
-
-    let grid = build_grid(&all_rows);
-    let col_count = grid.iter().map(Vec::len).max().unwrap_or(0);
-    if col_count == 0 {
-        return;
-    }
-
-    // Compute per-column alignment (first explicit cell wins) and widths.
-    let aligns = compute_column_aligns(&grid, col_count);
-    let widths = compute_column_widths(&grid, col_count, &aligns);
-
-    let header_count = table.header_rows.len();
-    for (idx, slots) in grid.iter().enumerate() {
-        serializer.write_line(&format_grid_row(slots, &widths, col_count));
-        // Separator row sits between the header rows and the body.
-        if header_count > 0 && idx + 1 == header_count {
-            serializer.write_line(&format_separator_row(&widths, &aligns));
-        }
-    }
-}
-
-fn compute_column_aligns(grid: &[Vec<Slot>], col_count: usize) -> Vec<TableCellAlignment> {
-    let mut aligns = vec![TableCellAlignment::None; col_count];
-    for slots in grid {
-        for (col, slot) in slots.iter().enumerate() {
-            if let Slot::Origin(cell) = slot {
-                if aligns[col] == TableCellAlignment::None && cell.align != TableCellAlignment::None
-                {
-                    aligns[col] = cell.align;
-                }
-            }
-        }
-    }
-    aligns
-}
-
-fn compute_column_widths(
-    grid: &[Vec<Slot>],
-    col_count: usize,
-    aligns: &[TableCellAlignment],
-) -> Vec<usize> {
-    let mut widths = vec![0usize; col_count];
-    for slots in grid {
-        for (col, slot) in slots.iter().enumerate() {
-            widths[col] = widths[col].max(slot.text().chars().count());
-        }
-    }
-    // Separator widths need a minimum of 3 (`---`) plus 1 for each
-    // colon a `:left`/`right:`/`:center:` marker adds. Round up so the
-    // separator row's `---` segment is at least as wide as the content.
-    for (col, w) in widths.iter_mut().enumerate() {
-        let min = match aligns.get(col).copied().unwrap_or(TableCellAlignment::None) {
-            TableCellAlignment::Center => 5,                           // `:---:`
-            TableCellAlignment::Left | TableCellAlignment::Right => 4, // `:---` / `---:`
-            TableCellAlignment::None => 3,
-        };
-        if *w < min {
-            *w = min;
-        }
-    }
-    widths
-}
-
-/// Append one `| text |`-style cell padded to `width` to `out` (the leading
-/// `|` of the row, and the `|` after each cell, are emitted here).
-fn push_cell(out: &mut String, text: &str, width: usize) {
-    out.push(' ');
-    out.push_str(text);
-    for _ in text.chars().count()..width {
-        out.push(' ');
-    }
-    out.push(' ');
-    out.push('|');
-}
-
-fn format_grid_row(slots: &[Slot], widths: &[usize], col_count: usize) -> String {
-    let mut out = String::from("|");
-    for (col, &width) in widths.iter().enumerate().take(col_count) {
-        // Columns past this (short) row's slots render as empty padding.
-        let text = slots.get(col).map(Slot::text).unwrap_or("");
-        push_cell(&mut out, text, width);
-    }
-    out
-}
-
-fn format_separator_row(widths: &[usize], aligns: &[TableCellAlignment]) -> String {
-    let mut out = String::from("|");
-    for (i, &w) in widths.iter().enumerate() {
-        out.push(' ');
-        let align = aligns.get(i).copied().unwrap_or(TableCellAlignment::None);
-        match align {
-            TableCellAlignment::Left => {
-                out.push(':');
-                for _ in 1..w {
-                    out.push('-');
-                }
-            }
-            TableCellAlignment::Right => {
-                for _ in 0..w.saturating_sub(1) {
-                    out.push('-');
-                }
-                out.push(':');
-            }
-            TableCellAlignment::Center => {
-                out.push(':');
-                for _ in 0..w.saturating_sub(2) {
-                    out.push('-');
-                }
-                out.push(':');
-            }
-            TableCellAlignment::None => {
-                for _ in 0..w {
-                    out.push('-');
-                }
-            }
-        }
-        out.push(' ');
-        out.push('|');
-    }
-    out
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::format::Format;
-    use lex_core::lex::ast::text_content::TextContent;
-    use lex_core::lex::testing::lexplore::{ElementType, Lexplore};
-    use lex_core::lex::testing::text_diff::assert_text_eq;
-
-    /// Drive the bare `LexSerializer` directly. NOTE: this bypasses
-    /// `LexFormat::serialize`'s pre/post passes (annotation inlining, blank
-    /// coalescing, trailing-blank trim), so its output can differ from
-    /// `lexd format`. Use `format_full` when the full pipeline matters (e.g.
-    /// annotation cases). Kept for the many element tests that assert the
-    /// serializer's raw structural output.
-    fn format_source(source: &str) -> String {
-        let format = super::super::LexFormat::default();
-        let doc = format.parse(source).unwrap();
-        let rules = FormattingRules::default();
-        let mut serializer = LexSerializer::new(rules);
-        doc.accept(&mut serializer);
-        serializer.output
-    }
-
-    /// Format through the full `LexFormat` pipeline (annotation inlining +
-    /// blank coalescing), i.e. what `lexd format` actually does — as opposed to
-    /// driving the bare `LexSerializer`. Needed for annotation cases, where the
-    /// pipeline strips the empty-paragraph marker artifact.
-    fn format_full(source: &str) -> String {
-        use crate::format::Format;
-        let format = super::super::LexFormat::default();
-        let doc = format.parse(source).unwrap();
-        format.serialize(&doc).unwrap()
-    }
-
-    // ==== Form-preserving roundtrip tests (#584 PR 3) =====================
-
-    #[test]
-    fn shortcut_form_round_trips_to_shortcut_spelling() {
-        // `:: author ::` source classifies as form=Shortcut for
-        // canonical `lex.metadata.author`. The formatter must emit the
-        // shortcut back, not the canonical. (The serializer's
-        // single-line-vs-block emission is a separate concern; this
-        // test focuses on the label-spelling preservation contract.)
-        let formatted = format_source(":: author :: Alice\n\nBody.\n");
-        assert!(
-            formatted.contains(":: author"),
-            "shortcut spelling should round-trip; got: {formatted}"
-        );
-        assert!(
-            !formatted.contains("lex.metadata.author"),
-            "canonical spelling must not leak into output: {formatted}"
-        );
-    }
-
-    #[test]
-    fn stripped_form_round_trips_to_stripped_spelling() {
-        // `:: metadata.category ::` classifies as Stripped — formatter
-        // must emit `metadata.category`, not the canonical.
-        let formatted = format_source(":: metadata.category :: tech\n\nBody.\n");
-        assert!(
-            formatted.contains(":: metadata.category"),
-            "stripped spelling should round-trip; got: {formatted}"
-        );
-        assert!(
-            !formatted.contains("lex.metadata.category"),
-            "canonical spelling must not leak: {formatted}"
-        );
-    }
-
-    #[test]
-    fn canonical_form_round_trips_unchanged() {
-        // `:: lex.metadata.title ::` classifies as Canonical and
-        // formats back as itself.
-        let formatted = format_source(":: lex.metadata.title :: My Doc\n\nBody.\n");
-        assert!(
-            formatted.contains(":: lex.metadata.title"),
-            "canonical spelling should round-trip; got: {formatted}"
-        );
-    }
-
-    #[test]
-    fn community_form_round_trips_unchanged() {
-        let formatted = format_source(":: acme.task id=42 :: foo\n\nBody.\n");
-        assert!(
-            formatted.contains(":: acme.task"),
-            "community label should round-trip; got: {formatted}"
-        );
-    }
-
-    #[test]
-    fn verbatim_shortcut_closer_round_trips() {
-        // `:: image src=x.png ::` (marker form) classifies as
-        // Shortcut for `lex.media.image`. The closing label must
-        // emit as `image`, not canonical.
-        let formatted = format_source("Photo subject:\n    alt text\n:: image src=\"x.png\" ::\n");
-        assert!(
-            formatted.contains(":: image"),
-            "verbatim closer should preserve shortcut: {formatted}"
-        );
-        assert!(
-            !formatted.contains("lex.media.image"),
-            "canonical must not leak: {formatted}"
-        );
-    }
-
-    // ==== Paragraph Tests ====
-
-    #[test]
-    fn test_paragraph_01_oneline() {
-        let source = Lexplore::load(ElementType::Paragraph, 1).source();
-        let formatted = format_source(&source);
-        assert_text_eq(
-            &formatted,
-            "This is a simple paragraph with just one line.\n",
-        );
-    }
-
-    #[test]
-    fn test_paragraph_02_multiline() {
-        let source = Lexplore::load(ElementType::Paragraph, 2).source();
-        let formatted = format_source(&source);
-        assert!(formatted.contains("This is a multi-line paragraph"));
-        assert!(formatted.contains("second line"));
-        assert!(formatted.contains("third line"));
-    }
-
-    #[test]
-    fn test_paragraph_03_special_chars() {
-        let source = Lexplore::load(ElementType::Paragraph, 3).source();
-        let formatted = format_source(&source);
-        assert!(formatted.contains("!@#$%^&*()"));
-    }
-
-    // ==== Session Tests ====
-
-    #[test]
-    fn test_session_01_simple() {
-        let source = Lexplore::load(ElementType::Session, 1).source();
-        let formatted = format_source(&source);
-        assert!(formatted.contains("Introduction\n"));
-        assert!(formatted.contains("    This is a simple session"));
-    }
-
-    #[test]
-    fn test_session_02_numbered_title() {
-        let source = Lexplore::load(ElementType::Session, 2).source();
-        let formatted = format_source(&source);
-        assert!(formatted.contains("1. Introduction:\n"));
-    }
-
-    #[test]
-    fn test_session_05_nested() {
-        let source = Lexplore::load(ElementType::Session, 5).source();
-        let formatted = format_source(&source);
-        // This is actually a complex doc with paragraphs and sessions
-        assert!(formatted.contains("1. Introduction {{session-title}}\n"));
-        assert!(formatted.contains("    This is the content of the session"));
-    }
-
-    // ==== List Tests ====
-
-    #[test]
-    fn test_list_01_dash() {
-        let source = Lexplore::load(ElementType::List, 1).source();
-        let formatted = format_source(&source);
-        assert!(formatted.contains("- First item\n"));
-        assert!(formatted.contains("- Second item\n"));
-    }
-
-    #[test]
-    fn test_list_02_numbered() {
-        let source = Lexplore::load(ElementType::List, 2).source();
-        let formatted = format_source(&source);
-        // Should normalize to sequential numbering
-        assert!(formatted.contains("1. "));
-        assert!(formatted.contains("2. "));
-        assert!(formatted.contains("3. "));
-    }
-
-    #[test]
-    fn test_list_03_alphabetical() {
-        let source = Lexplore::load(ElementType::List, 3).source();
-        let formatted = format_source(&source);
-        assert!(formatted.contains("a. "));
-        assert!(formatted.contains("b. "));
-        assert!(formatted.contains("c. "));
-    }
-
-    #[test]
-    fn test_list_04_mixed_markers() {
-        let source = Lexplore::load(ElementType::List, 4).source();
-        let formatted = format_source(&source);
-        // Should normalize to consistent markers
-        assert!(formatted.contains("1. First item\n"));
-        assert!(formatted.contains("2. Second item\n"));
-        assert!(formatted.contains("3. Third item\n"));
-    }
-
-    #[test]
-    fn test_list_07_nested_simple() {
-        let source = Lexplore::load(ElementType::List, 7).source();
-        let formatted = format_source(&source);
-        // Check for proper indentation of nested items
-        assert!(formatted.contains("- First outer item\n"));
-        assert!(formatted.contains("    - First nested item\n"));
-    }
-
-    #[test]
-    fn test_list_extended_markers_preserved() {
-        // NOTE: Extended markers (e.g., "1.2.3") require core parser support
-        // for Form::Extended. Currently the parser treats them as standard
-        // numbered lists, so normalization produces "1.", "2.", etc.
-        let source = "1.2.3 Item one\n1.2.4 Item two\n";
-        let formatted = format_source(source);
-        assert!(formatted.contains("1. Item one\n"));
-        assert!(formatted.contains("2. Item two\n"));
-    }
-
-    #[test]
-    fn test_list_extended_markers_nested_normalization() {
-        // Nested list with extended markers: formatter should rebuild hierarchical markers
-        let source = "Test:\n\n1. Outer level one\n    1.a Middle level one\n        1.a.1 Inner level one\n        1.a.2 Inner level two\n    1.b Middle level two\n2. Outer level two\n";
-        let formatted = format_source(source);
-        // Outer level items
-        assert!(
-            formatted.contains("1. Outer level one"),
-            "Expected '1. Outer level one' in: {formatted}"
-        );
-        assert!(
-            formatted.contains("2. Outer level two"),
-            "Expected '2. Outer level two' in: {formatted}"
-        );
-    }
-
-    #[test]
-    fn test_list_12_extended_form_fixture() {
-        let source = Lexplore::load(ElementType::List, 12).source();
-        let formatted = format_source(&source);
-        let formatted_again = format_source(&formatted);
-        assert_text_eq(&formatted, &formatted_again);
-    }
-
-    // ==== Definition Tests ====
-
-    #[test]
-    fn test_definition_01_simple() {
-        let source = Lexplore::load(ElementType::Definition, 1).source();
-        let formatted = format_source(&source);
-        assert!(formatted.contains("Cache:\n"));
-        assert!(formatted.contains("    Temporary storage"));
-    }
-
-    #[test]
-    fn test_definition_02_multi_paragraph() {
-        let source = Lexplore::load(ElementType::Definition, 2).source();
-        let formatted = format_source(&source);
-        // Should handle multiple paragraphs in definition body
-        assert!(formatted.contains("Microservice:\n"));
-        assert!(formatted.contains("    An architectural style"));
-        assert!(formatted.contains("    Each service is independently"));
-    }
-
-    // ==== Verbatim Tests ====
-
-    #[test]
-    fn test_verbatim_01_simple_code() {
-        let source = Lexplore::load(ElementType::Verbatim, 1).source();
-        let formatted = format_source(&source);
-        assert!(formatted.contains(":: javascript"));
-        assert!(formatted.contains("function hello()"));
-    }
-
-    #[test]
-    fn test_verbatim_02_with_caption() {
-        let source = Lexplore::load(ElementType::Verbatim, 2).source();
-        let formatted = format_source(&source);
-        // Should preserve verbatim content and captions
-        assert!(formatted.contains("API Response:"));
-    }
-
-    // ==== Annotation Tests ====
-
-    #[test]
-    fn test_annotation_01_marker_simple() {
-        let source = Lexplore::load(ElementType::Annotation, 1).source();
-        let formatted = format_full(&source);
-        // Marker annotation: closed `:: label ::` form (the open form is invalid
-        // and dropped on re-parse — lex#682).
-        assert_eq!(formatted, ":: note ::\n");
-    }
-
-    #[test]
-    fn test_annotation_02_with_params() {
-        let source = Lexplore::load(ElementType::Annotation, 2).source();
-        let formatted = format_full(&source);
-        assert_eq!(formatted, ":: warning severity=high ::\n");
-    }
-
-    #[test]
-    fn test_annotation_multi_param_keeps_comma_separator() {
-        // The parser uses the comma as the only parameter separator, so a
-        // space-only join collapses `k1=v1, k2=v2` into a single value on
-        // re-parse (lex#703). The formatted output must keep the comma and be a
-        // fixed point.
-        let formatted = format_full(":: warning type=critical, id=123 ::\n");
-        assert_eq!(formatted, ":: warning type=critical, id=123 ::\n");
-        assert_eq!(format_full(&formatted), formatted, "must be idempotent");
-    }
-
-    #[test]
-    fn test_annotation_05_block_paragraph() {
-        let source = Lexplore::load(ElementType::Annotation, 5).source();
-        let formatted = format_full(&source);
-        assert_eq!(
-            formatted,
-            ":: note ::\n    This is an important note that requires a detailed explanation.\n"
-        );
-    }
-
-    // ==== Round-trip Tests ====
-    // Format → parse → format should be idempotent
-
-    #[test]
-    fn test_round_trip_paragraph_01() {
-        let source = Lexplore::load(ElementType::Paragraph, 1).source();
-        let formatted = format_source(&source);
-        let formatted_again = format_source(&formatted);
-        assert_text_eq(&formatted, &formatted_again);
-    }
-
-    #[test]
-    fn test_round_trip_paragraph_02_multiline() {
-        let source = Lexplore::load(ElementType::Paragraph, 2).source();
-        let formatted = format_source(&source);
-        let formatted_again = format_source(&formatted);
-        assert_text_eq(&formatted, &formatted_again);
-    }
-
-    #[test]
-    fn test_round_trip_session_01() {
-        let source = Lexplore::load(ElementType::Session, 1).source();
-        let formatted = format_source(&source);
-        let formatted_again = format_source(&formatted);
-        assert_text_eq(&formatted, &formatted_again);
-    }
-
-    #[test]
-    fn test_round_trip_session_02_numbered() {
-        let source = Lexplore::load(ElementType::Session, 2).source();
-        let formatted = format_source(&source);
-        let formatted_again = format_source(&formatted);
-        assert_text_eq(&formatted, &formatted_again);
-    }
-
-    #[test]
-    fn test_round_trip_list_01_dash() {
-        let source = Lexplore::load(ElementType::List, 1).source();
-        let formatted = format_source(&source);
-        let formatted_again = format_source(&formatted);
-        assert_text_eq(&formatted, &formatted_again);
-    }
-
-    #[test]
-    fn test_round_trip_list_02_numbered() {
-        let source = Lexplore::load(ElementType::List, 2).source();
-        let formatted = format_source(&source);
-        let formatted_again = format_source(&formatted);
-        assert_text_eq(&formatted, &formatted_again);
-    }
-
-    #[test]
-    fn test_round_trip_list_03_alphabetical() {
-        let source = Lexplore::load(ElementType::List, 3).source();
-        let formatted = format_source(&source);
-        let formatted_again = format_source(&formatted);
-        assert_text_eq(&formatted, &formatted_again);
-    }
-
-    #[test]
-    fn test_round_trip_list_04_mixed_markers() {
-        let source = Lexplore::load(ElementType::List, 4).source();
-        let formatted = format_source(&source);
-        let formatted_again = format_source(&formatted);
-        assert_text_eq(&formatted, &formatted_again);
-    }
-
-    #[test]
-    fn test_round_trip_list_07_nested() {
-        let source = Lexplore::load(ElementType::List, 7).source();
-        let formatted = format_source(&source);
-        let formatted_again = format_source(&formatted);
-        assert_text_eq(&formatted, &formatted_again);
-    }
-
-    #[test]
-    fn test_round_trip_definition_01() {
-        let source = Lexplore::load(ElementType::Definition, 1).source();
-        let formatted = format_source(&source);
-        let formatted_again = format_source(&formatted);
-        assert_text_eq(&formatted, &formatted_again);
-    }
-
-    #[test]
-    fn test_round_trip_definition_02_multi() {
-        let source = Lexplore::load(ElementType::Definition, 2).source();
-        let formatted = format_source(&source);
-        let formatted_again = format_source(&formatted);
-        assert_text_eq(&formatted, &formatted_again);
-    }
-
-    #[test]
-    fn test_round_trip_verbatim_01() {
-        let source = Lexplore::load(ElementType::Verbatim, 1).source();
-        let formatted = format_source(&source);
-        let formatted_again = format_source(&formatted);
-        assert_text_eq(&formatted, &formatted_again);
-    }
-
-    #[test]
-    fn test_round_trip_verbatim_02_caption() {
-        let source = Lexplore::load(ElementType::Verbatim, 2).source();
-        let formatted = format_source(&source);
-        let formatted_again = format_source(&formatted);
-        assert_text_eq(&formatted, &formatted_again);
-    }
-
-    #[test]
-    fn test_verbatim_03_table_formatting() {
-        // PR 2 of #584 retired the legacy verbatim-with-markdown-body
-        // path: `:: doc.table ::` is forbidden, and `:: lex.tabular.table ::`
-        // / `:: tabular.table ::` source no longer round-trips through
-        // a markdown reformatter. The only surviving path is the
-        // structural Table element triggered by the bare `:: table ::`
-        // closer — `LexSerializer::visit_table` emits the pipe table
-        // directly with column alignment.
-        let source = "Table Example:\n    | A | B |\n    |---|---|\n    | 1 | 2 |\n:: table ::\n";
-        let formatted = format_source(source);
-
-        // Column-aligned pipe-table output from visit_table.
-        assert!(formatted.contains("| A   | B   |"));
-        assert!(formatted.contains("| --- | --- |"));
-        assert!(formatted.contains("| 1   | 2   |"));
-
-        // Also test with unformatted input — visit_table normalises.
-        let unformatted = "Table Example:\n    |A|B|\n    |-|-|\n    |1|2|\n:: table ::\n";
-        let formatted_2 = format_source(unformatted);
-
-        // Should be formatted nicely
-        assert!(formatted_2.contains("| A   | B   |"));
-        assert!(formatted_2.contains("| --- | --- |"));
-        assert!(formatted_2.contains("| 1   | 2   |"));
-    }
-
-    #[test]
-    fn test_round_trip_paragraph_then_verbatim_lex505() {
-        // Regression: Verbatim preceded by a paragraph must keep its leading
-        // blank line through a parse → serialize → parse round-trip. Without
-        // the blank, the re-parser merges the verbatim's subject line into
-        // the prior paragraph and the verbatim is lost. The parser consumes
-        // that blank as part of the verbatim's preamble (it doesn't appear
-        // as a BlankLineGroup in the AST), so the serializer has to emit it.
-        //
-        // Uses a `Title\n=====\n` header so the first line isn't absorbed as
-        // the document title — without it, "Intro paragraph." would become
-        // the doc title and the regression wouldn't be exercised.
-        let source =
-            "Doc\n===\n\nIntro paragraph.\n\nCode Example:\n\n    fn main() {}\n\n:: rust ::\n";
-        let formatted = format_source(source);
-        assert!(
-            formatted.contains("Intro paragraph.\n\nCode Example:"),
-            "expected blank line between paragraph and verbatim subject, got:\n{formatted}"
-        );
-        let formatted_again = format_source(&formatted);
-        assert_text_eq(&formatted, &formatted_again);
-    }
-
-    #[test]
-    fn test_verbatim_04_user_repro() {
-        // The original user input had dedented marker "::  doc.table ::"
-        // which caused parse-as-Definition + Document Annotation. The
-        // fix is to indent the marker to match the subject. Updated
-        // for PR 2 of #584: source uses the blessed `table` closer
-        // which triggers structural-Table parsing; the legacy verbatim
-        // path with markdown reformat is gone.
-        let source = "  The Table:\n    | Markup Language | Great |\n    |--------------------|--------|\n    | Markdown | No |\n    | Lex | Yes |\n  ::  table ::\n";
-
-        let formatted = format_source(source);
-
-        let table_start = formatted
-            .find("| Markup Language | Great |")
-            .expect("Table start not found");
-        let separator = formatted
-            .find("| --------------- | ----- |")
-            .expect("Separator not found");
-        // PR 3 of #584 wired form-preserving emission: the `:: table ::`
-        // source classifies as Shortcut, so the emitted closer is also
-        // `:: table ::`, not the canonical `:: lex.tabular.table ::`.
-        let footer_start = formatted.find(":: table ::").expect("Footer not found");
-
-        assert!(table_start < separator);
-        assert!(separator < footer_start);
-    }
-
-    #[test]
-    fn build_grid_pads_hole_before_trailing_rowspan() {
-        // Regression for the lex#694 review: a short continuation row whose cells
-        // run out before a rowspan-covered column further right must still emit
-        // that column's `^^` (and not let the carry leak into the next row).
-        // row0: a, b, c(rowspan 2) — c spans down into row1's third column.
-        // row1: a single cell `d`; the middle column is a hole, the third is `^^`.
-        let tc = |s: &str| TextContent::from_string(s.to_string(), None);
-        let row0 = TableRow::new(vec![
-            TableCell::new(tc("a")),
-            TableCell::new(tc("b")),
-            TableCell::new(tc("c")).with_span(1, 2),
-        ]);
-        let row1 = TableRow::new(vec![TableCell::new(tc("d"))]);
-        let grid = build_grid(&[&row0, &row1]);
-        let render = |slots: &[Slot]| {
-            slots
-                .iter()
-                .map(|s| s.text().to_string())
-                .collect::<Vec<_>>()
-        };
-        assert_eq!(render(&grid[0]), ["a", "b", "c"]);
-        assert_eq!(
-            render(&grid[1]),
-            ["d", "", "^^"],
-            "hole padded empty, trailing rowspan marker kept"
-        );
     }
 }
